+++ pom.xml
@@ -4,12 +4,12 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.collibra.interview</groupId>
-    <artifactId>interview-app</artifactId>
-    <version>1.0</version>
+    <groupId>com.collibra.backend.challenge</groupId>
+    <artifactId>backend-coding-challenge</artifactId>
+    <version>1.0-SNAPSHOT</version>
 
     <properties>
-        <main-class>com.collibra.interview.Application</main-class>
+        <main-class>com.collibra.backend.challenge.Application</main-class>
         <java.version>11</java.version>
         <encoding>UTF-8</encoding>
         <lombok.version>1.18.10</lombok.version>
