+++ src/main/java/com/collibra/backend/challenge/core/BaseMessageProcessor.java
@@ -1,0 +1,11 @@
+package com.collibra.backend.challenge.core;
+
+import com.collibra.backend.challenge.graph.DirectedGraph;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public abstract class BaseMessageProcessor implements MessageProcessor {
+
+    protected final DirectedGraph graph;
+
+}
