+++ src/main/java/com/collibra/backend/challenge/core/MessageProcessingException.java
@@ -1,0 +1,8 @@
+package com.collibra.backend.challenge.core;
+
+public abstract class MessageProcessingException extends RuntimeException {
+
+    public MessageProcessingException(final String message) {
+        super(message);
+    }
+}
