+++ src/main/java/com/collibra/backend/challenge/graph/Node.java
@@ -1,0 +1,14 @@
+package com.collibra.backend.challenge.graph;
+
+import lombok.Value;
+
+@Value
+public class Node implements Comparable<Node> {
+
+    private final String name;
+
+    @Override
+    public int compareTo(Node node) {
+        return name.compareTo(node.getName());
+    }
+}
