--- src/main/java/com/collibra/interview/core/MessageProcessingException.java
@@ -1,8 +1,0 @@
-package com.collibra.interview.core;
-
-public abstract class MessageProcessingException extends RuntimeException {
-
-    public MessageProcessingException(final String message) {
-        super(message);
-    }
-}
