--- src/main/java/com/collibra/interview/core/MessageProcessor.java
@@ -1,15 +1,0 @@
-package com.collibra.interview.core;
-
-import io.vavr.control.Option;
-
-public interface MessageProcessor {
-
-    boolean isApplicable(final String message);
-
-    String process(final String message);
-
-    default Option<String> getClientName() {
-        return Option.none();
-    }
-
-}
