--- src/main/java/com/collibra/interview/core/UnsupportedCommandException.java
@@ -1,11 +1,0 @@
-package com.collibra.interview.core;
-
-public class UnsupportedCommandException extends MessageProcessingException {
-
-    private static final String DEFAULT_MESSAGE = "SORRY, I DID NOT UNDERSTAND THAT";
-
-    public UnsupportedCommandException() {
-        super(DEFAULT_MESSAGE);
-    }
-
-}
